Add updateBoopFromSimulation utility and fix typos
