+++ packages/boop-sdk/lib/utils/getNonce.ts
@@ -1,30 +1,5 @@
 import { stringify } from "@happy.tech/common"
-
-export type GetNonceInput = {
-    /** Happy Account Address */
-    address: `0x${string}`
-    /** Nonce Track */
-    nonceTrack: bigint
-}
-
-export const GetNonce = {
-    Success: "getNonceSuccess",
-    Error: "getNonceError",
-} as const
-
-export type GetNonceStatus = (typeof GetNonce)[keyof typeof GetNonce]
-
-export type GetNonceOutput = GetNonceSuccess | GetNonceError
-
-export type GetNonceSuccess = GetNonceInput & {
-    status: typeof GetNonce.Success
-    nonceValue: bigint
-}
-
-export type GetNonceError = GetNonceInput & {
-    status: typeof GetNonce.Error
-    description: string
-}
+import { GetNonce, type GetNonceInput, type GetNonceOutput } from "../types/GetNonce"
 
 // "nonceValues(address,uint192)(uint64)"
 const nonceValuesSelector = "0xe631c8f2"
